Seed script now works
